Refactored for optional provider token
Some operations (like adding an initial admin user) require that
there not be a real token supplied to the client.

That being the case, we've made the `admin_token` field for the
provider itself optional, and we've added `creator_token` to the
`hephy_user` resource so the user's creator can be specified
per-user.